+++ src/common/services/recordNormalizing/recordProcessingCases.ts
@@ -1,14 +1,5 @@
-import {
-  BFLITE_URIS,
-  NEW_BF2_TO_BFLITE_MAPPING,
-  NON_BF_RECORD_ELEMENTS,
-} from '@common/constants/bibframeMapping.constants';
-
-export const getLabelUri = (blockKey: string, groupKey: string, fieldKey: string) => {
-  const typedMap = NEW_BF2_TO_BFLITE_MAPPING as BF2BFLiteMap;
-
-  return typedMap?.[blockKey]?.[groupKey]?.fields?.[fieldKey]?.label ?? '';
-};
+import { BFLITE_URIS } from '@common/constants/bibframeMapping.constants';
+import { getLookupLabelKey } from '@common/helpers/schema.helper';
 
 export const wrapWithContainer = (record: RecordEntry, blockKey: string, key: string, container: string) => {
   (record[blockKey][key] as unknown as string[]).forEach(recordEntry => {
@@ -24,7 +15,7 @@ export const wrapWithContainer = (record: RecordEntry, blockKey: string, key: st
 };
 
 export const wrapSimpleLookupData = (record: RecordEntry, blockKey: string, key: string) => {
-  const label = getLabelUri(blockKey, key, key);
+  const label = getLookupLabelKey(key);
 
   record[blockKey][key] = (record[blockKey][key] as unknown as string[]).map(recordEntry => ({
     [label]: [recordEntry],
@@ -32,18 +23,16 @@ export const wrapSimpleLookupData = (record: RecordEntry, blockKey: string, key:
 };
 
 export const notesMapping = (record: RecordEntry, blockKey: string) => {
-  const selector = NON_BF_RECORD_ELEMENTS[BFLITE_URIS.NOTE].container;
+  const selector = '_notes';
 
   if (!record[blockKey][selector]) return;
-
-  const label = getLabelUri(blockKey, selector, 'type');
 
   record[blockKey][selector] = (record[blockKey][selector] as unknown as RecordBasic[]).map(recordEntry => ({
     ...recordEntry,
     type: [
       {
         [BFLITE_URIS.LINK]: recordEntry.type,
-        [label]: [''],
+        [BFLITE_URIS.LABEL]: [''],
       },
     ],
   })) as unknown as RecursiveRecordSchema;
@@ -53,38 +42,6 @@ export const extractValue = (record: RecordEntry, blockKey: string, key: string,
   record[blockKey][key] = (record[blockKey][key] as unknown as RecordWithNestedFieldsDTO).map(
     recordEntry => recordEntry[source],
   ) as unknown as RecursiveRecordSchema;
-};
-
-export const processComplexGroupValues = (record: RecordEntry, blockKey: string, key: string, fieldName: string) => {
-  record[blockKey][key] = (record[blockKey][key] as unknown as RecordForComplexGroupsDTO).map(recordEntry => ({
-    [fieldName]: recordEntry,
-  })) as unknown as RecursiveRecordSchema;
-};
-
-export const processCreator = (record: RecordEntry, blockKey: string, key: string) => {
-  const selector = NON_BF_RECORD_ELEMENTS[BFLITE_URIS.CREATOR].container;
-  const label = getLabelUri(blockKey, key, selector);
-
-  record[blockKey][key] = (record[blockKey][key] as unknown as RecordProcessingDTO).map(recordEntry => {
-    const generatedValue = {
-      id: [recordEntry.id],
-      label: {
-        value: [recordEntry.label],
-        isPreferred: recordEntry.isPreferred,
-      },
-    } as unknown as Record<string, string[] | Record<string, string[]>[]>;
-
-    if (recordEntry[selector]) {
-      generatedValue[selector] = (recordEntry[selector] as unknown as string[])?.map((role: string) => ({
-        [BFLITE_URIS.LINK]: [role],
-        [label]: [''],
-      }));
-    }
-
-    return {
-      [recordEntry.type as unknown as string]: generatedValue,
-    };
-  }) as unknown as RecursiveRecordSchema;
 };
 
 export const processComplexLookup = (record: RecordEntry, blockKey: string, key: string) => {
@@ -110,17 +67,6 @@ export const processComplexLookup = (record: RecordEntry, blockKey: string, key:
 
     return generatedValue;
   }) as unknown as RecursiveRecordSchema;
-};
-
-export const processComplexGroupWithLookup = (
-  record: RecordEntry,
-  blockKey: string,
-  key: string,
-  fieldName: string,
-) => {
-  record[blockKey][key] = (record[blockKey][key] as unknown as string[]).map(recordEntry => ({
-    [fieldName]: recordEntry,
-  })) as unknown as RecursiveRecordSchema;
 };
 
 export const extractDropdownOption = (