+++ src/common/services/schema/index.ts
@@ -1,3 +1,2 @@
 export { SchemaWithDuplicatesService } from './schemaWithDuplicates.service';
-export { SchemaService } from './schema.service';
 export { SchemaGeneratorService } from './schemaGenerator.service';