+++ src/common/services/userValues/userValueTypes/userValueType.interface.ts
@@ -1,3 +1,3 @@
 export interface IUserValueType {
-  generate: ({ data, uri, uuid, labelSelector, uriSelector, type }: UserValueDTO) => UserValue | Promise<UserValue>;
+  generate: ({ data, uri, uuid, uriSelector, type }: UserValueDTO) => UserValue | Promise<UserValue>;
 }