+++ src/test/__tests__/common/services/recordToSchemaMapping/recordToSchemaMapping.service.test.ts
@@ -4,38 +4,12 @@ import { SchemaWithDuplicatesService } from '@common/services/schema';
 import { RecordToSchemaMappingService } from '@common/services/recordToSchemaMapping';
 import { getMockedImportedConstant } from '@src/test/__mocks__/common/constants/constants.mock';
 import * as BibframeMappingConstants from '@common/constants/bibframeMapping.constants';
-import * as FeatureConstants from '@common/constants/feature.constants';
 import { StatusType } from '@common/constants/status.constants';
 import { getLabelEntry, schema } from './data/schema.data';
 import { updatedSchema, updatedSchemaWithRepeatableSubcomponents } from './data/updatedSchema.data';
 import { mockInstanceTemplateMetadata, record, recordWithRepeatableSubcomponents } from './data/record.data';
 
-const mockedNewBf2ToBFLiteMapping = getMockedImportedConstant(BibframeMappingConstants, 'NEW_BF2_TO_BFLITE_MAPPING');
 const mockedBFLiteUris = getMockedImportedConstant(BibframeMappingConstants, 'BFLITE_URIS');
-const mockedCustomProfileEnabled = getMockedImportedConstant(FeatureConstants, 'CUSTOM_PROFILE_ENABLED');
-mockedNewBf2ToBFLiteMapping({
-  block_1: {
-    uriBFLite_literal_1: {
-      container: { bf2Uri: 'propertyURI_1' },
-      fields: {
-        uriBFLite_literal_1: { bf2Uri: 'propertyURI_1' },
-      },
-    },
-    uriBFLite_simple_1: {
-      container: { bf2Uri: 'bf2Uri_simple_1' },
-      fields: {
-        uriBFLite_simple_1: { bf2Uri: 'bf2Uri_simple_1' },
-      },
-    },
-    uriBFLite_group_1: {
-      container: { bf2Uri: 'propertyURI_2' },
-      options: { uriBFLite_option_1: { bf2Uri: 'dropdownOptionResourceURI_1' } },
-      fields: {
-        uriBFLite_option_literal_1: { bf2Uri: 'dropdownOption_1_PropertyURI_1' },
-      },
-    },
-  },
-});
 mockedBFLiteUris({ LINK: 'testLinkUri' });
 
 const recordBlocks: RecordBlocksList = ['block_1'];
@@ -171,9 +145,7 @@ describe('RecordToSchemaMappingService', () => {
       expect(selectedEntriesService.remove).toHaveBeenCalledWith('testKey-7');
     });
 
-    test('handles dropdown option when CUSTOM_PROFILE_ENABLED is true', async () => {
-      mockedCustomProfileEnabled(true);
-
+    test('handles dropdown option', async () => {
       const mockRecord = {
         block_1: {
           uriBFLite_group_1: [