+++ src/types/serviceContext.d.ts
@@ -7,7 +7,6 @@ type IRecordNormalizingService =
 type IRecordToSchemaMappingService =
   import('@common/services/recordToSchemaMapping/recordToSchemaMapping.interface').IRecordToSchemaMapping;
 type ISchemaService = import('@common/services/schema/schema.interface').ISchema;
-type IRecordGeneratorServiceLegacy = import('@common/services/record/record.interface').IRecordGenerator;
 type ISchemaGeneratorService = import('@common/services/schema/schemaGenerator.interface').ISchemaGenerator;
 type IRecordGeneratorService = import('@common/services/recordGenerator/recordGenerator.interface').IRecordGenerator;
 
@@ -18,8 +17,6 @@ type ServicesParams = {
   lookupCacheService?: ILookupCacheService;
   recordNormalizingService?: IRecordNormalizingService;
   recordToSchemaMappingService?: IRecordToSchemaMappingService;
-  schemaCreatorService?: ISchemaService;
-  recordGeneratorServiceLegacy?: IRecordGeneratorServiceLegacy;
   schemaGeneratorService?: ISchemaGeneratorService;
   recordGeneratorService?: IRecordGeneratorService;
 };